server_api/requirements.txt
fastapi
uvicorn
httpx

opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1

opentelemetry-exporter-otlp-proto-grpc==1.39.1

opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-instrumentation-logging==0.60b1